+++ python/qibuild/test/test_parser.py
@@ -1,0 +1,6 @@
+def test_guess_solve_deps(build_worktree):
+    world = build_worktree.create_project("world")
+    hello = build_worktree.create_project("hello", depends=["world"])
+
+    print build_worktree.build_projects
+